Avoid comparing signed and unsigned (VZ)
git-svn-id: svn://svn.sv.gnu.org/lmi/lmi/trunk@6531 8995fb05-a277-41b4-a4d2-e84fd280a9a6

+++ multidimgrid_any.cpp
@@ -1418,8 +1418,8 @@ void MultiDimAxisAnyChoice::UponSelectionChange(wxCommandEvent&)
 
 void MultiDimAxisAnyChoice::SelectionChanged()
 {
-    unsigned int const sel = GetSelection();
-    if(!(0 <= sel && sel < axis_.GetCardinality()))
+    int const sel = GetSelection();
+    if(!(0 <= sel && sel < static_cast<int>(axis_.GetCardinality())))
         {
         fatal_error()
             << "The axis is inconsistent with its choice control."
